Add nix based docker package install and runtime
This allows specifying the version of docker to install and run.
Startup is slow because the version specified is installed every time,
but this guarantees that the configured version is what is run.  It's
a little horky to support vresions older than 'latest', but the nix
package hash for the version needs to be defined in docker.versions so
that exact version can be pulled in.